20240821
this is minor update is for lightPublish:

- change the version number according to LightPublish version, to according to the original the \light\images\vXXXX

Future
======
It is now using os.path.split(os.path.sep) to split the vesion number, which it is potentially not accurate if the N drive folder structure changed in the future. Should try to update to toolkit, easier to update in the future.